feat(wallet): widen aggregate wallet capacity to $2,000,000
The wallet balance grew far beyond math.MaxInt32 raw quota
(~$4,294 at QuotaPerUnit=500000), making every Stripe and Waffo
Pancake amount request fail ValidateTopUpQuotaCapacity with
"top-up quota limit exceeded" and display a pending amount of 0.

Introduce an explicit aggregate wallet ceiling MaxWalletQuota =
1_000_000_000_000 ($2,000,000) and migrate only the user-wallet
domain (users.quota, its readers, caches, and update paths) to
int64, while keeping common.MaxQuota = math.MaxInt32 for the
per-request charge / log / per-top-up entitlement domains.

All four settlement paths now credit through creditTopUpQuota(),
an atomic conditional UPDATE (WHERE quota <= cap - credited,
RowsAffected==1 as sole success signal across SQLite/MySQL/
PostgreSQL) that fails closed with ErrTopUpQuotaLimitExceeded or
ErrInvalidTopUpQuota. Epay and Stripe settlement propagate the
sentinel; Creem/Waffo/WaffoPancake wrap it into their generic
top-up failure with rollback to Pending.

+++ controller/topup_stripe_test.go
@@ -462,7 +462,7 @@ func insertStripeWebhookTopUp(t *testing.T, db *gorm.DB, tradeNo string, session
 	return userID
 }
 
-func getStripeWebhookUserQuota(t *testing.T, db *gorm.DB, userID int) int {
+func getStripeWebhookUserQuota(t *testing.T, db *gorm.DB, userID int) int64 {
 	t.Helper()
 	var user model.User
 	require.NoError(t, db.Select("quota").Where("id = ?", userID).First(&user).Error)
@@ -483,7 +483,7 @@ func TestStripeWebhook_MatchingBindingRecoversSessionAndSettlesExactlyOnce(t *te
 	require.NotNil(t, success)
 	assert.Equal(t, common.TopUpStatusSuccess, success.Status)
 	assert.Equal(t, sessionID, success.ExpectedSessionID)
-	expectedQuota := int(2 * common.QuotaPerUnit)
+	expectedQuota := int64(2 * common.QuotaPerUnit)
 	assert.Equal(t, expectedQuota, getStripeWebhookUserQuota(t, db, userID))
 
 	replay := invokeStripeWebhook(signedStripeWebhookRequest(t, payload))
